requirements.txt
web3>=6.0.0
python-dotenv>=1.0.0
requests>=2.28.0
tqdm>=4.65.0
